Fixed 'deploy' make target with custom driver image
If hostname of the image contains port, and semicolon, deploy script
will try to split on it, and create an incorrect kustomize yaml.
Should only split on the semicolon after the last slash, and if tag does
not exist, substitute it with 'latest'.
Also, fix some things 'shellcheck' is unhappy about.