NO-JIRA: Clarify Personas for Sharing
Add language to describe who is typically responsible for sharing
resources in a cluster. Application developers granted the "edit" role
in a namespace may not have sufficient permissions to create the RBAC
needed to consume a shared resource. This would be left to a platform
engineer or other system administrator. Adding this language helps
clarify potential friction points when using the CSI driver.

Signed-off-by: Adam Kaplan <adam.kaplan@redhat.com>